requirements.txt
faiss-cpu
numpy
torch
transformers
requests
beautifulsoup4
gradio